Refactor global scripts to not expose global variables.
