+++ src/_includes/scripts/firebase.njk
@@ -14,4 +14,4 @@
   // Initialize Firebase
   firebase.initializeApp(firebaseConfig);
 </script>
-<script src="/scripts/helpers/auth.js"></script>
+<script src="/scripts/helpers/globals.mjs" type="module"></script>
